[SP-297] Remove old Alembic versions, create fresh revision
