--- rdl/alembic/versions/0d4a3ce9c0a9_add_failure_reason_column_to_data_load_.py
@@ -1,29 +1,0 @@
-"""add failure_reason column to data_load_execution
-
-Revision ID: 0d4a3ce9c0a9
-Revises: 710e28aa5978
-Create Date: 2019-04-03 12:15:16.898526
-
-"""
-from alembic import op
-import sqlalchemy as sa
-
-
-# revision identifiers, used by Alembic.
-revision = '0d4a3ce9c0a9'
-down_revision = '710e28aa5978'
-branch_labels = None
-depends_on = None
-
-
-def upgrade():
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.add_column('data_load_execution', sa.Column('failure_reason',
-                                                   sa.String(length=1000), nullable=True), schema='rdl')
-    # ### end Alembic commands ###
-
-
-def downgrade():
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.drop_column('data_load_execution', 'failure_reason', schema='rdl')
-    # ### end Alembic commands ###
