--- rdl/alembic/versions/710e28aa5978_add_data_load_execution_table.py
@@ -1,50 +1,0 @@
-"""add data_load_execution table
-
-Revision ID: 710e28aa5978
-Revises:
-Create Date: 2019-04-03 11:52:21.994634
-
-"""
-from alembic import op
-import sqlalchemy as sa
-from sqlalchemy.dialects import postgresql
-
-# revision identifiers, used by Alembic.
-revision = "710e28aa5978"
-down_revision = None
-branch_labels = None
-depends_on = None
-
-
-def upgrade():
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.execute("CREATE SCHEMA IF NOT EXISTS rdl")
-    op.create_table(
-        "data_load_execution",
-        # sa.Column("id", sa.INTEGER(), nullable=True),
-        sa.Column("execution_id", sa.String(length=250), nullable=False),
-        sa.Column("model_name", sa.String(length=250), nullable=False),
-        sa.Column("status", sa.String(length=25), nullable=False),
-        sa.Column("last_sync_version", sa.BigInteger(), nullable=False),
-        sa.Column("sync_version", sa.BigInteger(), nullable=False),
-        sa.Column("is_full_refresh", sa.Boolean(), nullable=False),
-        sa.Column("full_refresh_reason", sa.String(length=100), nullable=False),
-        sa.Column(
-            "completed_on",
-            sa.DateTime(timezone=True),
-            server_default=sa.text("now()"),
-            nullable=True,
-        ),
-        sa.Column("execution_time_ms", sa.Integer(), nullable=False),
-        sa.Column("rows_processed", sa.Integer(), nullable=False),
-        sa.Column("model_checksum", sa.String(length=100), nullable=False),
-        # sa.PrimaryKeyConstraint("id"),
-        schema="rdl",
-    )
-    # ### end Alembic commands ###
-
-
-def downgrade():
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.drop_table("data_load_execution", schema="rdl")
-    # ### end Alembic commands ###
