+++ rdl/alembic/versions/b62653121263_initial_revision.py
@@ -1,0 +1,74 @@
+"""Initial revision
+
+Revision ID: b62653121263
+Revises: 
+Create Date: 2019-08-22 13:58:14.585775
+
+"""
+from alembic import op
+import sqlalchemy as sa
+
+
+# revision identifiers, used by Alembic.
+revision = "b62653121263"
+down_revision = None
+branch_labels = None
+depends_on = None
+
+
+def upgrade():
+
+    # Generated using `alembic -c rdl/alembic.ini -x $DESTINATION_DB_URL revision -m "$REVISION_MESSAGE" --autogenerate`
+    op.execute("CREATE SCHEMA IF NOT EXISTS rdl")
+
+    # ### commands auto generated by Alembic - please adjust! ###
+    op.create_table(
+        "execution",
+        sa.Column("execution_id", sa.String(length=250), nullable=False),
+        sa.Column("created_on", sa.DateTime(timezone=True), nullable=False),
+        sa.Column("updated_on", sa.DateTime(timezone=True), nullable=False),
+        sa.Column(
+            "status", sa.String(length=50), server_default="Started", nullable=False
+        ),
+        sa.Column("started_on", sa.DateTime(timezone=True), nullable=False),
+        sa.Column("completed_on", sa.DateTime(timezone=True), nullable=True),
+        sa.Column("execution_time_s", sa.BigInteger(), nullable=True),
+        sa.Column("rows_processed", sa.BigInteger(), nullable=True),
+        sa.Column("batches_processed", sa.Integer(), nullable=True),
+        sa.Column("models_processed", sa.Integer(), nullable=True),
+        sa.PrimaryKeyConstraint("execution_id"),
+        schema="rdl",
+    )
+    op.create_table(
+        "execution_model",
+        sa.Column("execution_model_id", sa.String(length=250), nullable=False),
+        sa.Column("created_on", sa.DateTime(timezone=True), nullable=False),
+        sa.Column("updated_on", sa.DateTime(timezone=True), nullable=False),
+        sa.Column("execution_id", sa.String(length=250), nullable=False),
+        sa.Column("model_name", sa.String(length=250), nullable=False),
+        sa.Column(
+            "status", sa.String(length=50), server_default="Started", nullable=False
+        ),
+        sa.Column("last_sync_version", sa.BigInteger(), nullable=False),
+        sa.Column("sync_version", sa.BigInteger(), nullable=False),
+        sa.Column("is_full_refresh", sa.Boolean(), nullable=False),
+        sa.Column("full_refresh_reason", sa.String(length=100), nullable=False),
+        sa.Column("started_on", sa.DateTime(timezone=True), nullable=False),
+        sa.Column("completed_on", sa.DateTime(timezone=True), nullable=True),
+        sa.Column("execution_time_ms", sa.BigInteger(), nullable=True),
+        sa.Column("rows_processed", sa.BigInteger(), nullable=True),
+        sa.Column("batches_processed", sa.Integer(), nullable=True),
+        sa.Column("model_checksum", sa.String(length=100), nullable=False),
+        sa.Column("failure_reason", sa.String(length=1000), nullable=True),
+        sa.ForeignKeyConstraint(["execution_id"], ["rdl.execution.execution_id"]),
+        sa.PrimaryKeyConstraint("execution_model_id"),
+        schema="rdl",
+    )
+    # ### end Alembic commands ###
+
+
+def downgrade():
+    # ### commands auto generated by Alembic - please adjust! ###
+    op.drop_table("execution_model", schema="rdl")
+    op.drop_table("execution", schema="rdl")
+    # ### end Alembic commands ###
